requirements.txt
python-dotenv
pandas
sendgrid
# plotly express